# src/main/resources/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Framework Spark",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4567"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Listar todos",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user": {
      "post": {
        "summary": "Cadastrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/user/{cpf}": {
      "get": {
        "summary": "Buscar por CPF",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Atualizar",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "Deletar",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "cpf": { "type": "string" },
          "phone": { "type": "string" },
          "street": { "type": "string" },
          "number": { "type": "string" },
          "neighborhood": { "type": "string" },
          "zipCode": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" }
        }
      }
    }
  }
}